Error handling and bug fixing.
Polishing / fixing all functionality before we can merge into develop.

+++ src/js/stores/node.js
@@ -26,21 +26,37 @@ export default Reflux.createStore({
     this.trigger(node)
   },
 
-  // On Remove will remove the node itself (RDF file) and disconnects it from the center node.
+  /**
+   * @summary Deletes a rdf file and it's connection to the center node
+   * and plays the delete animation
+   * @param {object} node - the node to be deleted.
+   * @param {object} centerNode - we dissconnect from this node.
+   */
+
   onRemove(node, centerNode){
     let subject = rdf.sym(centerNode.uri)
     let predicate = rdf.sym(node.connection)
     let object = rdf.sym(node.uri)
 
-    this.gAgent.deleteFile(object.uri).then(()=>{
-      // Deleting the connection to the file. 
-      this.gAgent.deleteTriple(subject, predicate, object).then(()=>{
-        graphActions.deleteNode(node) 
-      // Basic error handling
-      }).catch((e)=>{console.log('Error', e ,'while removing connection')})
-    }).catch((e)=>{console.log('error', e, 'occured while deleting')})
+    this.gAgent.deleteFile(object.uri).then((response)=>{
+      if (response.ok) {
+        this.gAgent.deleteTriple(subject, predicate, object).then((result)=>{
+          if (result.ok)
+            graphActions.deleteNode(node) 
+        })
+      } else { 
+        console.log('An error occured while removing the node.')
+      }
+    })
 	}, 
 
+
+  /**
+   * @summary Disconnects a node from another node.
+   * @param {object} subject - triple subject describing connection
+   * @param {object} predicate - triple predicate describing connection
+   * @param {object} object - triple object describing connection
+   */
 
   onDisconnectNode(node, centerNode){
     let subject = rdf.sym(centerNode.uri)
@@ -48,6 +64,15 @@ export default Reflux.createStore({
     let object = rdf.sym(node.uri)
     this.gAgent.deleteTriple(subject, predicate, object)
 	},
+
+
+  /**
+   * @summary Links a node to another node.
+   * @param {string} start - subject uri describing connection
+   * @param {string} type - predicate uri describing connection
+   * @param {string} end - object uri describing connection
+   * @param {boolean} flag - fire the animation in the graph 
+   */
 
   link(start, type, end, flag) {
     let predicate = null
